+++ src/renderer/src/modules/media/media-metadata/index.ts
@@ -1,0 +1,1 @@
+export { default as MediaMetadata } from './media-metadata';